moved components to ui folder (shadcn structure)
