+++ packages/create-llama/templates/streaming/nextjs/app/components/ui/chat-avatar.tsx
